# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Savezy API
  version: 1.0.0
  description: |
    REST API for Savezy, providing authentication workflows and expense management
    for authenticated users. All expense endpoints require a JWT Bearer token
    issued by the authentication flow.
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.savezy.com
    description: Production server (example)
...
paths:
  /api/auth/google/init:
    get:
      tags: [Auth]
      summary: Start Google OAuth flow
      description: Returns the Google OAuth authorization URL for the mobile client.
      parameters:
        - name: redirect_uri
          in: query
          required: true
          description: Mobile redirect URI registered in the backend configuration.
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Authorization URL generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleInitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/google/callback:
    post:
      tags: [Auth]
      summary: Exchange Google authorization code for JWT
      description: |
        Completes the Google OAuth authorization flow. Expects a one-time state token,
        authorization code, and original mobile redirect URI.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleCallbackRequest'
      responses:
        '200':
          description: JWT issued successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/google/verify:
    post:
      tags: [Auth]
      summary: Verify Google ID token from client SDKs
      description: Verifies a Google ID token obtained directly from a mobile client and issues a JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleVerifyRequest'
      responses:
        '200':
          description: JWT issued successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/token/verify:
    post:
      tags: [Auth]
      summary: Verify JWT token validity
      description: Checks whether a JWT access token is valid and returns its payload.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: JWT token to verify (if not supplied in Authorization header).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GoogleInitResponse:
      type: object
      properties:
        auth_url:
          type: string
          format: uri
          example: https://accounts.google.com/o/oauth2/v2/auth?...state=abc
    GoogleCallbackRequest:
      type: object
      required: [code, state, redirect_uri]
      properties:
        code:
          type: string
        state:
          type: string
        redirect_uri:
          type: string
          format: uri
    GoogleVerifyRequest:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string
          description: Google ID token obtained from the client SDK.
    AuthSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT access token issued by the backend.
        user:
          $ref: '#/components/schemas/User'
    VerifyTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        payload:
          type: object
          properties:
            user_id:
              type: integer
            email:
              type: string
              format: email
            iat:
              type: integer
              description: Issued at timestamp (epoch seconds).
            exp:
              type: integer
              description: Expiration timestamp (epoch seconds).
        error:
          type: string
          nullable: true
          description: Present when the token is invalid.
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        picture:
          type: string
          format: uri
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Validation failed.
        data:
          type: object
          additionalProperties: true